requirements.txt
black==19.10b0
eth-brownie>=1.11.0,<2.0.0
dotmap==1.3.23